+++ src/app/components/form-base/form-base.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormGroup } from '@angular/forms';
 
-// TODO: adapt location, plant and photos to use this
 @Component({
   selector: 'ltm-form-base',
   standalone: true,