+++ src/app/components/plant/buttons/plant-button-water/plant-button-water.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { booleanAttribute, Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
@@ -16,5 +16,5 @@ import {
   styleUrls: ['./plant-button-water.component.scss'],
 })
 export class PlantButtonWaterComponent extends PlantBaseActionComponent {
-  @Input() disabled: boolean = false;
+  @Input({ transform: booleanAttribute }) disabled: boolean = false;
 }