+++ src/app/components/plant/forms/plant-form-description/plant-form-description.component.ts
@@ -20,12 +20,14 @@ import { FormBaseComponent } from '@components/form-base/form-base.component';
   templateUrl: './plant-form-description.component.html',
 })
 export class PlantFormDescriptionComponent implements FormBaseComponent {
-  @Input() currentDescription: string | null = '';
+  @Input() currentDescription?: string | null;
   form = this.fb.group({ description: [''] });
 
   constructor(private fb: FormBuilder) {}
 
   ngOnInit(): void {
-    this.form.patchValue({ description: this.currentDescription });
+    if (this.currentDescription) {
+      this.form.patchValue({ description: this.currentDescription });
+    }
   }
 }