+++ src/app/components/user/user-edit/user-edit.component.ts
@@ -49,7 +49,6 @@ import { AuthService } from '@services/auth.service';
 import { ErrorHandlerService } from '@services/error-handler.service';
 import { User } from '@models/user.model';
 
-// TODO: remove photo
 @Component({
   selector: 'ltm-user-edit',
   standalone: true,
@@ -73,6 +72,7 @@ export class UserEditComponent {
   @ViewChildren('form') formComponents!: QueryList<FormBaseComponent>;
   @ViewChild('formEmail') emailComponent!: UserFormEmailComponent;
   @ViewChild('formUsername') usernameComponent!: UserFormUsernameComponent;
+  @ViewChild('fileUploader') fileUploaderComponent!: FileUploaderComponent;
 
   userForm: FormGroup = this.fb.group({
     username: new FormControl<string>('', Validators.required),
@@ -150,7 +150,10 @@ export class UserEditComponent {
     const wd = this.openWaitDialog();
 
     if (user) {
-      this.api.editUser(user).pipe(
+      const removePicture =
+        !!this.fileUploaderComponent.form.get('remove')?.value;
+
+      this.api.editUser(user, removePicture).pipe(
         finalize(() => { wd.close() }),
         catchError((err: HttpErrorResponse) => {
           const error = err.error;